Added ability for the `has*()` and `get*()` methods to accept delimited strings that contain a series of keys in one string.
The delimiters can be configured through the constructor, they default to the characters "\", "/" & ".".

When fetching a key from the config, you can mix & match arrays, arguments lists, and delimited strings.

Example:
```
$config->getKey( 'level1.level2', 'level3', [ 'level4\level5', 'level6' ] );
```

This allows you to collect keys from everywhere and not worry about the format in which you get them. The only thing that matters is the order of the keys as they appear.

See #7